pyproject.toml
[project]
name = "froide_payment"
readme = "README.md"
description = ""
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities",
]
version = "0.0.3"
dependencies = [
  "django-payments[stripe]",
  "django-countries",
  "django-prices>=2.1.0",
  "django-localflavor",
  "python-dateutil",
  "prices",
  "stripe>=2.29.4",
  "requests",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["froide_payment*"]

[tool.djlint]
ignore = "T002,T003,H005,H006,H021,H023,H029,H030,H031"
